requirements.txt
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.0
greenlet==1.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
python-dotenv==0.21.0
SQLAlchemy==1.4.41
flask_migrate==2.7.0
Werkzeug==2.2.2
zipp==3.8.1